pyproject.toml
[project]
name = "pyakta"
version = "0.1.0"
description = "A pure Python implementation of ZCAP-LD (Authorization Capabilities for Linked Data)"
authors = [
    {name = "Luke Hinds", email = "lukehinds@gmail.com"},
]


dependencies = [
    "base58>=2.1.1",
    "pynacl>=1.5.0",
    "pydantic>=2.11.0",
    "pyld>=2.0.4",
]


requires-python = ">=3.11"
readme = "README.md"
license = {text = "Apache License 2.0"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=pyakta"

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "ruff~= 0.9",
    "bandit>=1.8.3",
]

[project.urls]
"Homepage" = "https://github.com/lukehinds/pyakta"
"Repository" = "https://github.com/lukehinds/pyakta"